refactor getWeakSubjectivityCheckpoint
