+++ src/renderer/ducks/beacon/sagas.ts
@@ -63,11 +63,7 @@ import {
     readBeaconChainNetwork,
 } from "../../services/eth2/client/module";
 import {getClientParams} from "../../services/docker/getClientParams";
-import {WeakSubjectivityCheckpoint} from "../../components/ConfigureBeaconNode/ConfigureBeaconNode";
-import {HttpClient} from "../../services/api";
-import {getNetworkConfig} from "../../services/eth2/networks";
-import {IBeaconConfig} from "@chainsafe/lodestar-config";
-import cheerio from "cheerio";
+import {getWeakSubjectivityCheckpoint} from "./getWeakSubjectivityCheckpoint";
 
 export function* pullDockerImage(
     image: string,
@@ -97,66 +93,6 @@ export function* pullDockerImage(
         yield put(endDockerImagePull());
         return false;
     }
-}
-
-type BeaconScanWSC = {
-    // eslint-disable-next-line camelcase
-    current_epoch: number;
-    // eslint-disable-next-line camelcase
-    is_safe: boolean;
-    // eslint-disable-next-line camelcase
-    ws_checkpoint: string;
-    // eslint-disable-next-line camelcase
-    ws_period: number;
-};
-
-function* getWeakSubjectivityCheckpoint(
-    type: WeakSubjectivityCheckpoint,
-    meta: string,
-    network: string,
-    // eslint-disable-next-line camelcase
-): Generator<CallEffect | Promise<string>, string, BeaconScanWSC & string & {current_finalized_epoch: number}> {
-    if (type !== WeakSubjectivityCheckpoint.none) {
-        if (type === WeakSubjectivityCheckpoint.custom) return meta;
-        if (type === WeakSubjectivityCheckpoint.infura) {
-            const config = getNetworkConfig(network);
-            const api = new CgEth2ApiClient((config as unknown) as IBeaconConfig, meta);
-            return yield call(api.beacon.state.getWeakSubjectivityCheckpoint);
-        }
-        if (type === WeakSubjectivityCheckpoint.beaconScan) {
-            const httpClient = new HttpClient(`https://beaconscan.com/`);
-            if (network === "mainnet") {
-                const ws = yield httpClient.get(`ws_checkpoint`);
-                if (ws) return `${ws.ws_checkpoint}:${ws.current_epoch}`;
-            }
-
-            // TODO: change scraper with api after etherscan implement it
-            const dom = yield httpClient.get(network);
-            const home = cheerio.load(dom);
-            const href = home("#finalizedSlot a")[0].attribs["href"];
-            const domSlot = yield httpClient.get(href);
-            const slot = cheerio.load(domSlot);
-            const root = slot("#ContentPlaceHolder1_divDetail > div:nth-child(2) > div.col-md-9.font-size-1").text();
-            const epoch = slot(
-                "#overview > div > div > div:nth-child(2) > div.col-md-9.js-focus-state.font-size-1 > a",
-            ).text();
-            return `${root}:${epoch}`;
-        }
-        if (type === WeakSubjectivityCheckpoint.beaconChain) {
-            // TODO: change scraper with api after beaconcha.in implement it
-            const httpClient = new HttpClient(`https://${network !== "mainnet" ? network + "." : ""}beaconcha.in/`);
-            // eslint-disable-next-line camelcase
-            const result = yield httpClient.get("/index/data");
-            const dom = yield httpClient.get(`/epoch/${result.current_finalized_epoch}`);
-            const home = cheerio.load(dom);
-            const roots: string[] = [];
-            home("tbody i").each((_, el) => {
-                roots.push(el.attribs["data-clipboard-text"]);
-            });
-            return `${roots.reverse()[0]}:${result.current_finalized_epoch}`;
-        }
-    }
-    return "";
 }
 
 function* startLocalBeaconSaga({
